+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Status of the `main` branch. Changes prior to the next official version change w
 * General:
   * Various fixes related to indexing, special paths and determation of ignored paths
   * Decreased `TOOL_DEFAULT_MAX_ANSWER_LENGTH` to be in accordance with (below) typical max-tokens configurations
+  * Allow passing language server specific settings through `ls_specific_settings` field (in `serena_config.yml`)
 
 # 0.1.4